Flytter arbeidsforholdene i egen mappe, pga flere komponenter i dette panelet relatert til arbeidsforhold. Koblet opp korrekte felter og omrigg av elementer.
