+++ src/felles-komponenter/arbeidsforhold/arbeidsforholdet.less
@@ -1,4 +1,4 @@
-@import '../utils/mixins';
+@import '../../utils/mixins';
 
 .arbeidsforholdet {
 
@@ -7,7 +7,7 @@
   }
 
   .arbeidsforholdet__detaljer {
-    .beskrivliste-mixin('row');
+    .beskrivliste-mixin('col');
   }
 
   .arbeidsavtale {
